tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animationevent_init_bubble.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - canBubbleArg in initAnimationEvent</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-events-">
   <meta name="flags" content="dom">
   <meta name="assert" content="canBubbleArg in initAnimationEvent specifies
   whether or not the event can bubble">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var divs = document.querySelectorAll('div');
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = true;
         divs[0].addEventListener('animationend', function() {
            passed = false;
         });
         setTimeout(function() {
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS' : 'FAIL');
         }, 500);
         try {
            var evt = document.createEvent('AnimationEvent');
            evt.initAnimationEvent('animationend', false, false, 'test', 0);
            divs[1].dispatchEvent(evt);
         } catch(e) {
            passed = false;
         }
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
   <div>
      <div></div>
   </div>
</body>
</html>
